requirements.txt
pytest
playwright
allure-pytest
pyyaml
selenium
